backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0.post1
gunicorn==21.2.0

# Firebase + Auth
firebase-admin==6.5.0

# Google Cloud SDKs
google-cloud-firestore==2.16.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.20.0

# Redis Cache
redis==5.0.1

# Pydantic + Env management
pydantic==2.6.1
pydantic-settings==2.2.1

# Upload parsing
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
